Move ResourceType enum into Empire/ResourcePool.h
