+++ Empire/ResourcePool.cpp
@@ -1,12 +1,51 @@
 #include "ResourcePool.h"
 
+#include <cassert>
 #include <boost/lexical_cast.hpp>
-
+#include "../universe/Enums.h"
 #include "../universe/ObjectMap.h"
 #include "../universe/Planet.h"
-#include "../universe/Enums.h"
 #include "../util/AppInterface.h"
 #include "../util/Logger.h"
+
+MeterType ResourceToMeter(ResourceType type) {
+    switch (type) {
+    case RE_INDUSTRY:   return METER_INDUSTRY;
+    case RE_RESEARCH:   return METER_RESEARCH;
+    case RE_INFLUENCE:  return METER_INFLUENCE;
+    case RE_STOCKPILE:  return METER_STOCKPILE;
+    default:
+        assert(0);
+        return INVALID_METER_TYPE;
+        break;
+    }
+}
+
+MeterType ResourceToTargetMeter(ResourceType type) {
+    switch (type) {
+    case RE_INDUSTRY:   return METER_TARGET_INDUSTRY;
+    case RE_RESEARCH:   return METER_TARGET_RESEARCH;
+    case RE_INFLUENCE:  return METER_TARGET_INFLUENCE;
+    case RE_STOCKPILE:  return METER_MAX_STOCKPILE;
+    default:
+        assert(0);
+        return INVALID_METER_TYPE;
+        break;
+    }
+}
+
+ResourceType MeterToResource(MeterType type) {
+    switch (type) {
+    case METER_INDUSTRY:    return RE_INDUSTRY;
+    case METER_RESEARCH:    return RE_RESEARCH;
+    case METER_INFLUENCE:   return RE_INFLUENCE;
+    case METER_STOCKPILE:   return RE_STOCKPILE;
+    default:
+        assert(0);
+        return INVALID_RESOURCE_TYPE;
+        break;
+    }
+}
 
 //////////////////////////////////////////////////
 // ResourcePool
