+++ universe/Enums.cpp
@@ -1,46 +1,5 @@
 #include "Enums.h"
 
-#include <cassert>
-
-
-MeterType ResourceToMeter(ResourceType type) {
-    switch (type) {
-    case RE_INDUSTRY:   return METER_INDUSTRY;
-    case RE_RESEARCH:   return METER_RESEARCH;
-    case RE_INFLUENCE:  return METER_INFLUENCE;
-    case RE_STOCKPILE:  return METER_STOCKPILE;
-    default:
-        assert(0);
-        return INVALID_METER_TYPE;
-        break;
-    }
-}
-
-MeterType ResourceToTargetMeter(ResourceType type) {
-    switch (type) {
-    case RE_INDUSTRY:   return METER_TARGET_INDUSTRY;
-    case RE_RESEARCH:   return METER_TARGET_RESEARCH;
-    case RE_INFLUENCE:  return METER_TARGET_INFLUENCE;
-    case RE_STOCKPILE:  return METER_MAX_STOCKPILE;
-    default:
-        assert(0);
-        return INVALID_METER_TYPE;
-        break;
-    }
-}
-
-ResourceType MeterToResource(MeterType type) {
-    switch (type) {
-    case METER_INDUSTRY:    return RE_INDUSTRY;
-    case METER_RESEARCH:    return RE_RESEARCH;
-    case METER_INFLUENCE:   return RE_INFLUENCE;
-    case METER_STOCKPILE:   return RE_STOCKPILE;
-    default:
-        assert(0);
-        return INVALID_RESOURCE_TYPE;
-        break;
-    }
-}
 
 const std::map<MeterType, MeterType>& AssociatedMeterTypes() {
     static const std::map<MeterType, MeterType> meters = {
